Remodelling how shell expressions run in shells
Summary:
This CL introduces the new `Runner` trait that serves two purposes:
1. Increase code readability by getting rid of the functional `run_in_shell` interface that was used to execute shell expressions in a bash process. It grew beyond it's original intent and became increasingly complex.
2. Prepare for the new style `Executor` (next Diffs), that requires a tight integration with `bash`. The `Runner` interface leaves room for additional implementations of other shells (zsh etc) without requiring them to be done prematurely while ensuring clear responsibility boundaries.

With that in mind the `ThreadedRunner` struct now provides what the former functional implementation was providing (i.e. starting an arbitrary process, piping the shell expression into it, gathering the output and return it while upholding timeout constraints).

This `Runner` is used by the `BashScriptExecutor` (formerly `SequentialExecutor`), that was renamed to better describe the underlying implementation (all executions are run within a single bash script).

Lastly, the restructuring made it necessary to move out the `reaper` code - soon to be entirely removed likely (D47335933)

Reviewed By: abesto

Differential Revision: D47229178

fbshipit-source-id: f456ebd4d728ba47ed90f9f163082d17255ebdac