Stateful Isolated Executor
Summary:
This CL introduces the `StatefulExecutor` which and the `StatefulBashRunner`. In conjunction they
- Run each execution in an individual shell process, which makes sure that
  - The output streams of individual executions can be captured in isolation from other executions
  - Timeouts and other per-test configuration can be applied, well, per test
- Run all executions within a consistent state, as if they were all executed from within the same shell (bash) process

The here provided implementation of the runner is bound to `bash`, but the designs allows for adding support for additional shells (`StatefulZshRunner` etc) in the future.
The implementation also leaves room to work with detached processes (e.g. starting a server, then waiting for it) and generally prepares for the upcoming configuration syntax.

There is one notable difference to the behavior of the previous `BashScriptExecutor`: Each execution implicitly has `shopt -s expand_aliases` set. Without that alias support would be broken. A future configuration could be introduced to disable that behavior. This is also the reason why the `BashScriptExecutor` is still in place and used for Cram test execution.

## Under the hood

The `StatefulExecutor` creates new instances of the `StatefulBashRunner` for each sequential execution. Each runner instance uses the same, shared temporary directory. The bash `trap` is then used to persist environment variables, shell variables, aliases, functions and other "bash state" in a file `<state-directory>/state` after each execution. This `state` file is then sourced by each following execution, which then restores the state of the previous execution.

## Risks

While the implementation has been tested against all `scrut_test` in Meta, there could be relevant state that has been overlooked and hence is not captured and restored. Or worse: that cannot be captured or restored.

Reviewed By: abesto

Differential Revision: D47229176

fbshipit-source-id: 2f09c181ff719144928b5fb0c35ef5bc061fe5b3