Document behavior of detached process that use standard streams
Summary:
This CL adds an integration test that documents the Scrut's behavior when a process that is detached (e.g. `command &` or `nohup comand &`) when reading from `/dev/stdin` or writing to `/dev/stdout` or `/dev/stderr`, namely: The rust execution that runs the bash process that executes the shell expression of the test will wait for the standard streams to close, meaning the detaching will not work.

Since this behavior is a) very edge-case and b) soon to change (with configuration) this is not added to documentation.

Reviewed By: uepoch

Differential Revision: D48439655

fbshipit-source-id: 4bd2e47f593b3482827542fcbed24b8929d3205a