Replace execution logic with subprocess crate
Summary:
This CL simplifies the code that executes commands from within Scrut by using the 3rd party `subprocess` crate instead of using the low level standard library interface.

## How does it work?

Scrut executes a bash process for each shell expression in a test file.

This bash process was formerly started using rust standard library primitives:
- A `std::process::Command` of `bash` was spawned in a thread from `std::thread`
- In another thread `STDIN` (i.e. the test shell expression) was written into the command
- In additional threads `STDERR` and `STDOUT` where collected
- If a timeout for the execution was used then another thread was started that killed the execution thread (above) after timeout

All this code and thread handling is gone as the `subprocess` crate comes with an interface that keeps threads out of our code, reducing the complexity.

The downside is that we are loosing low-level process control - for which there is no use-case so far.

Reviewed By: AndreasBackx

Differential Revision: D47335933

fbshipit-source-id: ea6e2a97ec0318d3bc9996457d308115ca59f185