Fix Windows `exit` execution handling in subset of Cram tests
Summary:
This CL fixes test execution result handling in Windows when users explicitly call `exit`.

# What was failing?

Before this CL Cram tests failed to delegate explicitly set exit codes correctly in certain conditions (observed and reproducible): test files that contained eight or more test expressions).

For example:

```
A cram test
  $ echo something
  $ exit 80
  $ echo weirdly
  $ echo only
  $ echo in
  $ echo cases
  $ echo of
  $ echo more than 8 tests in the file
```

The above example would not result in a skip (`exit 80` tells Scrut to skip), but would result in an execution error.

# How does the fix work?

Underlying the subprocess crate is used to execute shell expressions in a `bash` process. The interface of that crate is comparable to the Python subprocess library: Popen + communicate. When this communicate call fails in Windows, this CL now also waits for the process exit code explicitly and uses that to determine the exit code.

I suspect the issue to be caused in the [somewhere here](https://github.com/hniksic/rust-subprocess/blob/release/0.2.7/src/communicate.rs#L382-L412) in the subprocess crate.

Reviewed By: AndreasBackx

Differential Revision: D49318816

fbshipit-source-id: a02343bb3f77eb33cdb130cb370993fb05994ac3