Parser returns testcases and document configuration
Summary:
This CL changes the `Parser::parse` signature to also return the `DocumentConfig` in addition to the list of `TestCase`s. Right now there is no actually parsing done within the parser implementations, but a default (empty) configuration is returned.

This step is in preparation of doing the actually parsing and done so to make review easier, because this touched a lot of code-lines.

The biggest change here is adding all the new configuration related structures into the Python adapter.

Reviewed By: AndreasBackx

Differential Revision: D50320538

fbshipit-source-id: 2817b41ec0de34069f79397d34bbd677477c1b8c