Add (per) document and (per) testcase configuration
Summary:
This CL adds configuration per test file document and per testcase within, as agreed upon in the [RFC](https://docs.google.com/document/d/1_1vGRNLiyzEQP2q8gLyZ97EaclEeMVKThWUgeBKfuXs/edit#heading=h.mbq89nxz23tq).

In order to keep the review simple, no use of the configuration is made.

In a following Diff I will extend the `TestCase` struct to contain the`TestCaseConfig` as an attribute and respective extended the `Parser::parse` return signature to also contain the `DocumentConfig` in addition to the list of test cases. This would touch a lot of files (as a lot of files have test that use `TestCase`) and would detract from the thing I want to be reviewed: The structure of the configuration clases.

Reviewed By: AndreasBackx

Differential Revision: D50120449

fbshipit-source-id: 17474cd2191b20e46c346df049c24a7221a1128e