Store configuration with testcase
Summary:
This CL extends the `TestCase` struct with a new attribute `config` that stores the `TestCaseConfig` for this particular testcase.

All code-points where testcases are used (especially test) where updated in a way that the `config` attribute contains an empty (default) value.

This step is a precursor to:
- Extending the `Parser` to read the provided configuration from the Markdown files
- Utilize the newly provided configuration during execution and test validation

**Why not not `Option<TestCase>` in `config` of `TestCase`?**

Every `TestCase` will contain configuration. This configuration will be different by the source of the `TestCase` alone, e.g. a `TestCase` extracted from a cram file will have `keep_crlf` set to `true` whereas the same `TestCase` parsed from a markdown file will be `false` (unless it is set differently by the test author, of course).

The information of the source-type (markdown / cram) is lost after parsing, as it is of no consequence. Hence every `TestCase` will cary all information (configuration) that is needed to validate it with itself (barring the `total_timeout` attribute from `DocumentConfig` that pertains to all testcases).

There will be no `TestCase` instance with a null value.

Reviewed By: AndreasBackx

Differential Revision: D50222441

fbshipit-source-id: 7fe9621757b1528e68e6ade440ab778556e25c73