Parse and render per-document and per-testcase configuration
Summary:
This CL enables parsing of inline per-document and per-testcase configuration and rendering of (updated) markdown documents with configuration.

# What was done?

The Markdown parser is extended to
- Parse any YAML front-matter into document configuration.
- Parse any YAML per-testcase configuration that is appended to the language (after the three backticks that denote the code block) and append to test the testcase

That means that Scrut is now aware of the provided configuration.

# How does the behavior change

There is no functional change.

- Configuration is parsed and internally stored **but not used**
- Configuration is kept in `scrut update` operations

# Caveat

Per-testcase is currently only supported when provided as one-liner.

# Alternatives

Parsing Markdown documents is currently done in a line-by-line fashion. Since we are not really parsing any markdown, but instead just extracting code blocks that are marked with `scrut` + some smol things, this has not been a problem. However, it feels not nice whenever I touch it.

Now would be again another time to rethink that and use a "proper" Markdown parser - although we might now breaking (some) Markdown, as the per-testcase configuration is ... at least not standard.

Rewriting this would likely consume much more time than this little addendum in this Diff. Also the `scrut create/update` document generation would of course be affected.

Please opine, if you have experience with specific Markdown parsers and would like to volunteer some time :)

Reviewed By: AndreasBackx

Differential Revision: D50322379

fbshipit-source-id: a1f9a38f55880fbe0db21559a7c708254cd2cf86