Extract configuration from command lines
Summary:
This CL adds methods for the create, test and update commands, that convert the user-provided command line parameters into per-testcase and per-document configuration, that can be merged with the parsed per-testcase and per-document configuration (in a later Diff).

Right now this does not change the behaviour of Scrut, as the configurations are only extracted from the command line, but not used.

Reviewed By: AndreasBackx

Differential Revision: D50598009

fbshipit-source-id: e297269de080a086d0c86236980e8ce2e519de84