Parsers return testcases with parsed config
Summary:
The CL ensures that both the markdown and the cram parser return test cases that contain the inline configuration.

# What does it do?

For Cram that means that the returned testcases contain the cram default configuration, as there is no support for inline configuration.

For Markdown that means that the returned testcases contain the markdown default configuration, amended by per-document and per-testcase configuration that the user provides.

# What is the impact?

The only behavioral change is that invalid testcase or document configuration would trigger an error. As configuration is not yet in use, this should not have any behavioral impact.

This change prepares for the biggest step in this chain: making use of the configuration.

Reviewed By: AndreasBackx

Differential Revision: D50598475

fbshipit-source-id: 407da8f9dc795bd23e614fd87513fa640c9fea13