Remove markdown language configuration from document config
Summary:
Remove the (so far unusable) `language_markers` attribute from per-document configuration and hide the `--markdown-languages` parameter.

## Why?

This would require bigger refactoring in test and update command, due to how test files are being parsed currently.

Right now the only user of the command line parameter, and we should / will remove that ourselves (follow up diffs).

TL;DR: Lots of effort now, for feature that turns out not useful (nobody uses it, we won't soon either).

## What changed

This does not change any behavior, as line configuration is not enabled as of now. The parameter for the command line parameter (`--markdown-languages`) still works the same.

Reviewed By: uepoch

Differential Revision: D50650514

fbshipit-source-id: c0a051000cd9db87c437e9fa185ce12975ceac56