Enable per-document configuration: defaults
Summary:
This CL enables the `defaults` configuration per-document.

# What changed?

Defaults for all test cases in the same test file can now be provided using the per-document `defaults` configuration like so:

```
 ---
defaults:
  timeout: 1m
 ---

Every test in this file would have a timeout of 1 minute

... %< .. tests follow here .. %< ..
```

This affects the order of precedence of configuration like so:
1. Command line provided (if applicable)
2. Per-TestCase (if given)
3. Default in Per-Document (if given)
4. Default

# Why?

If all (or most) testcases in a file share the same (non-default) configuration, it is much easier (to read and maintain) to have it noted only once.

Reviewed By: AndreasBackx

Differential Revision: D51424910

fbshipit-source-id: f17e77bb55c87186aa14b404f9ebdf1572c0bb3e