Enable per-testcase configuration: skip_document_code
Summary:
This CL renames the per-testcase configuration attribute that specifies which exit code tells Scrut to skip the whole document - and enables the feature.

# What changes?

The `skip_code` configuration was renamed to `skip_document_code`. As this configuration was not enabled before hand, it does not change existing tests.

The major change is that this feature is now enabled and users can change the default skip document code 80 to a custom value (see added testcases).

# Why?

Purpose: This feature allows users to write test files that are only for specific conditions (e.g. OS-specific, etc).

Rename: The new name makes it clearer that, if a test issues this exit code, the whole document is skipped and not just the test.

Reviewed By: AndreasBackx

Differential Revision: D51205007

fbshipit-source-id: 4febd18a43d770ef7332efbc0fcb590984194824