Chore: make work and temporary directory mandatory
Summary:
This CL makes the work and temporary directory (that are provided for each test file execution) non-optional

# What changed?

Each test file execution via an `Executor` requires two parameters:
1. A list of testcases from that were parsed from the test file
2. A context, that contains the per-document configuration, a work directory (which us usually a temporary directory, but can also provided by the user) and a temporary directory.

The two directories used to be optional (`Option<PathBuf>`) and are now not anymore (`PathBuf`).

As the context is part of many unit tests (that pertain to test execution), making those directories non-optional required a test-setup, that maintains temporary directories for individual unit test executions. Adding this constitutes the majority of the changed (added) lines.

# Why?

In the (distant) past the directories could have been omitted. By now there is no scenario where they are not  expected to be present.

The `Executor` implementations needed to contain unnecessary code that accounted for the directories not being provided, which was never the case, aside from testing. Hence: tests did use the production code paths. This was bad. It had to go.

Reviewed By: AndreasBackx

Differential Revision: D51394957

fbshipit-source-id: 97f9485958593434620bc697c4c0dd263d64cb70