Prepare total timeout global parameter
Summary:
This CL prepares the document-wide (total) timeout for the upcoming inline configuration.

# What changed?

Behavioral Scrut acts still the same.

Under the hood the `--timeout-seconds` command line parameter - that is available in `test`, `update` and `create` - is now a global parameter.

Both executors are now prepared to use the (already) default per-document execution time of 900 seconds, in case the delegated per-document timeout has no value (`None`).

# Why?

When inline-configuration will be activated, then the following order of precedence must be observed:
1. Parameter first: If `--timeout-seconds` is provided by the user then this value must be used in all test documents.
2. Inline-configuration second: If no command line parameter is provided, then whatever is specified in the per-document configuration must be used
3. Default last: If neither of the above is provided, then the default (900 seconds) must be used.

Currently the command line parameter is non-optional, that means it always has a value, which would always override (2).

This CL prepares for making the command line parameter optional when inline configuration is activated (following in Diff stack)

Reviewed By: AndreasBackx

Differential Revision: D51202362

fbshipit-source-id: a6b2289c3b275c75ed531977367b2e38a5668183