Enable per-testcase configuration: wait
Summary:
This CL enables the the `wait` per-testcase configuration.

# What changed?

When a test is configured to wait, Scrut will delay it's execution until either the time runs or or a optionally specified file shows up. For example:

  scrut {wait: 1m}

Or

  scrut {wait: {timeout: 1m, path: some-file}}

The path `some-file` is expected to be created in the `$TEMPDIR` that is accessible and consistent for all tests in the same test file.

# Why?

This makes server/client tests feasible, where a preceding test that is `{detached: true}` starts a server (and possibly creates a file once the server is started) and a following test runs a client that communicates with that server. Or comparable.

Reviewed By: AndreasBackx

Differential Revision: D51394953

fbshipit-source-id: 7cd75e9bbbf16d7f749b7b16f2f7a3d7cd31572f