Fix test for detached execution in bash runner
Summary:
This CL fixes the detached execution test in bash runner.

# What changes?

Scrut provides each execution of a shell expression of a testcase an execution context, which comes with a temp and a working directory. During unit testing, this execution context is setup to create those directories - and remove them when the context is dropped.

A newly added test (which tested that detached testcases do not persist state when using the bash runner) failed, because the drop of the context happened while the (detached) exuction was still running (and accessing the directory).

This CL makes sure hat the context lives longer than the execution, which fixes the test.

Reviewed By: uepoch

Differential Revision: D51560337

fbshipit-source-id: 0a92d22bf3de48a3077ce1d021b58f50cd795992