Fix read cram config
Summary:
This CL fixes a Cram config edge case that occurs when multiple testcases that are separated by empty lines are provided.

# What changed

Before this CL the following test would have confused the Cram parse:

```
  $ true
  $ true

  $ true
```

The 2nd testcase would have been parsed without a config, which would have resulted in an error like the following, when compared against the 3rd testcase during parseing:

```
2023-12-11T15:39:05.665178Z ERROR cli_log: failing in "clifoundation/scrut/selftest/cram/parse-configs.t": error in execution with index=1: inconsistent configuration value for keep_crlf
```

This CL fixes that.

Reviewed By: abesto

Differential Revision: D52036332

fbshipit-source-id: 756853c91856c6ee4023f1c9e6aaed7a9bfe6eb2