Print missing no-eol in pretty test result
Summary:
This CL ensures Scrut suggest to add `(no-eol)` with the pretty renderer if it is missing.

# What?

Append ` (no-eol)` to unexpected output lines that do not end in a new-line in the pretty validation result renderer

# Why?

Before this CL, if a test output validation contained an `equal` rule (i.e. a comparison with a line that ends in a new-line) but expected a `no-eol` rule (i.e. a comparison with a line that does not end with a new-line), then the pretty rendere would provide:

```
1     | - There is no new line
   1  | + There is no new line
```

This was unclear and did not provide the test author with any actionable steps to fix the test. With this CL the output will now correctly show:

```
1     | - There is no new line
   1  | + There is no new line (no-eol)
```

Reviewed By: AndreasBackx

Differential Revision: D63902398

fbshipit-source-id: 0703165dc80b41e1fe32a837493ae320baf5f759