Add Github Action to build and release binaries of new tags
Summary:
This adds a Github Action build and release binaries for Linux, MacOS and Windows upon new new tag.

# Why?

Simplify usage of Scrut by making binaries available for common platforms.

# What?

Add Github Action that starts automatic build and release for the following targets:
- x86_64-unknown-linux-musl
- aarch64-unknown-linux-musl
- x86_64-apple-darwin
- aarch64-apple-darwin
- x86_64-pc-windows-gnu

The binaries will be build, archived (zip for windows, tar zst for others) and then uploaded as a a release asset for a newly created release version.

Reviewed By: AndreasBackx

Differential Revision: D68104600

fbshipit-source-id: 35a781e96aa33dcefcd43a9f4357b0c3dbfd7de5