Show output on timeout
Summary:
This CL makes timeouts being treated as validation errors.

# Why?

So far Scrut treated timeouts in test executions as fatal errors, which resulted in Scrut ending all test execution and printing out the file where the timeout occurred (see below).

For trivial cases where only few test-cases where in a single test document and where timeouts occur consistently this is enough to identify the issue.

However, if the document contains a large amount of of test cases, or when timeouts occur only occasionally, or when timeouts only occur in hard to reproduce environments (think windows test infra), then debugging such timeouts is hard and time consuming.

# What?

Treat timeouts as validation errors, which gives:
- Timeout in one test document does not stop the execution of other test documents
- Timeout error points to the specific test-case in the document where the timeout happened
- Timeout error provides STDOUT and STDERR of the timed out execution (as much as is printed by the executed command before it is aborted)

This makes it much easier for users to identify
1. Which command (test case) timed out
2. Why the timeout occurred (assuming the executed command is verbose enough)

Reviewed By: AndreasBackx

Differential Revision: D71402437

fbshipit-source-id: 0a7406f08754ed69f85c2afdebf2dcac24cfed8a