Installation script
Summary:
The CL adds an `install.sh` script that can be used to install OSS Scrut on Mac/Linux.

# Why?

Provide an additional option to install Scrut in OSS usage, to make lower the barrier of using Scrut.

# What?

Add `install.sh` that downloads the current / latest version of Scrut from GH releases and installs it in user's `$HOME/bin` or `$HOME/.local/bin` directory (or user-definable location).

Reviewed By: AndreasBackx

Differential Revision: D70324805

fbshipit-source-id: 17850bff1ff812aabccc6aaa66e471748504e5a4