Align namings with documentation
Summary:
This CL uses "document" and "testcase" wording in results.

# Why?

Tests that are executed by Scrut are stored in files. This makes the word "test" ambiguous (e.g. "Writing a test" can mean: *Adding a test to a file* or *creating a new file with one or multiple tests within*).

To make it clear what is meant in which situation Scrut uses "documents" (or "test documents") to refer to the files themselves and "testcases" to refer to the individual tests within a file.

The documentation already accounts for this differentiation - for most part (and will fully after the current rewrite is done).

# What?

The result line that is printed after `scrut test` execution is changed from before:

```
Result: 1 file(s) with 1 test(s): 1 succeeded, 0 failed and 0 skipped
```

To now:

```
Result: 1 document(s) with 1 testcase(s): 1 succeeded, 0 failed and 0 skipped
```

# Note

Most of the changes here are test files where the `Result:..` line is printed and validated.

Primary review-relevant contents in `src/**/*.rs`.

Note also that most of the `info!` and `debug!` statements will be replaced further down this diff stack (so don't care about the actual content they are printing)

Reviewed By: abesto

Differential Revision: D71500385

fbshipit-source-id: 500b1fc38f80671412796560327b9fe78caa82d5