Add `strip_ansi_escaping` configuration
Summary:
This CL adds a per-test configuration option to remove ANSI escape sequences before applying output validation.

# Why?

Testing rich CLIs that use color to improve the user-experience is hard in Scrut, as  tests that contain escape sequences in outputs are hard to read and reason about. The color coding is not essential to validate whether a CLI is working. Stripping such escape sequences greately simplifies test writing and maintenance.

# What?

- Add `strip_ansi_escaping` configuration, that defaults to `false`
- If (per-testcase) set to `true` then all ANSI sequences are stripped from the output before validation is applied

Reviewed By: AndreasBackx

Differential Revision: D71749611

fbshipit-source-id: 16bb2a5229e8246bc004899183adebde52076ca7