Allow for empty line continuations
Summary:
When creating a file with empty lines with a heredoc, a trailing space is currently required so that the continuation is accepted. This is bad because text editors and lint rules automatically strip trailing spaces—they're also invisible in most editors, making problems hard to diagnose. This diff makes it so that continuations work if there is no trailing space and the line is empty.

See D77661218 and D78051355 for real world examples

Reviewed By: ukautz

Differential Revision: D78321181

fbshipit-source-id: 63d9370c2ab09a272040f2cd3ed5b3019fc0a21c