escape lines before calculating max line length
Summary:
We were calculating the max line length of all the lines. Then we were looping over the lines and altering it, making them longer as we were non-printable characters. Then we were doing `max - line.length()` which would underflow because `line.length()` would be longer on the longest line.

This changes it so we're escaping all lines first.

Reviewed By: richardjrossiii

Differential Revision: D88066881

fbshipit-source-id: 33dac68a37caced18b4a302fc27072d2c2116811